Improve CI to upload dev artefacts Rename port_null to port_template to build with CI
