+++ .github/workflows/build.yml
@@ -8,6 +8,7 @@ on:
   push:
     branches:
       - master
+      - dev*
 
 jobs:
   build:
@@ -16,6 +17,9 @@ jobs:
       BUILD_TAG: 0
 
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        board: [teensy41,template]
 
     steps:
     - name: Checkout repo
@@ -31,7 +35,7 @@ jobs:
         git submodule init hw/mcu/nxp/mcux-sdk
         git submodule update
         cd ../../../
- 
+
     - name: Set up Python
       uses: actions/setup-python@v1
 
@@ -43,9 +47,22 @@ jobs:
         echo "BUILD_TAG=$BUILD_TAG" >> $GITHUB_ENV
         
     - name: Compile code
-      run: platformio run
+      run: |
+        platformio platform update
+        platformio run -e ${{ matrix.board }}
+        # platformio system prune --force
 
-    - if: github.event_name == 'push' 
+    - name: Upload firmware artifact to current run
+      uses: actions/upload-artifact@v2
+      with:
+        name: test-firmware-${{ matrix.board }}
+        path: |
+          .pio/build/**/*.hex
+          .pio/build/**/*.bin
+          .pio/build/**/*.elf
+          .pio/build/**/*.map
+
+    - if: github.event_name == 'push' && github.ref == 'refs/heads/master' #TODO handle previews
       name: Create Release
       id: create_release
       uses: "actions/create-release@v1"
@@ -55,12 +72,12 @@ jobs:
         draft: false
         prerelease: false
 
-    - if: github.event_name == 'push' 
-      name: Upload binary to release
+    - if: github.event_name == 'push' && github.ref == 'refs/heads/master' #TODO handle previews
+      name: Upload binary for release
       id: upload-release-asset 
       uses: actions/upload-release-asset@v1
       with:
           upload_url: ${{ steps.create_release.outputs.upload_url }}
-          asset_path: .pio/build/teensy41/firmware.hex
-          asset_name: firmware-teensy41.hex
+          asset_path: .pio/build/${{ matrix.board }}/firmware.hex
+          asset_name: firmware-${{ matrix.board }}.hex
           asset_content_type: application/hex
