+++ platformio.ini
@@ -124,17 +124,17 @@ build_flags =
     -DUSE_HAL_DRIVER
     -DSTM32F750xx
 
-[env:null]
+[env:template]
 platform = native
 
 src_filter =
     ${common_env_data.src_filter}
-    +<port_null/*.cpp>
-    +<port_null/*.c>
+    +<port_template/*.cpp>
+    +<port_template/*.c>
     +<lib/tinyusb/src/portable/template/*.c>
 
 build_flags =
     ${common_env_data.build_flags}
-    -Isrc/port_null/
+    -Isrc/port_template/
     -DCFG_TUSB_MCU=OPT_MCU_NONE
     -lpthread
